+++ docs/releases.rst
@@ -1,1 +1,3 @@
+:tocdepth: 1
+
 .. include:: ../RELEASES.rst